Update TwitterSearchAlgorithm2.java
Sort array3 in descending order after we have created the HashMap from its values and the variable from our for loop. The loop that creates our HashMap only needs to be as long as our num variable because thats all we are outputting. Change long b = 0; to int b = 0; for faster output.