Update TwitterSearchAlgorithm.java
Used a sort method to find which values from array3 are the next highest value found and sorted in descending order. This is one of the ways we can write this program. And we only need to make array4 int[] array because we reassign the value until we find which why y from our inner loop in our sort is the last highest position in line in our sort algorithm. Keeping track of this position is one way to keep track of account number instead of a HashMap. This is the way to some this problem, and only requires 4 arrays.